refactor: create Repository test
AnswerRepositoryTest.java
DeleteHistoryRepositoryTest.java
