+++ src/test/java/qna/domain/DeleteHistoryRepositoryTest.java
@@ -1,0 +1,78 @@
+package qna.domain;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+
+import java.util.List;
+import java.util.Optional;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+import static qna.domain.DeleteHistoryTest.DH_ANSWER;
+import static qna.domain.DeleteHistoryTest.DH_QUESTION;
+
+@DataJpaTest
+public class DeleteHistoryRepositoryTest {
+    @Autowired
+    private DeleteHistoryRepository deleteHistoryRepository;
+
+    private DeleteHistory answerSavedDeleteHistory;
+    private DeleteHistory questionDeleteHistory;
+
+    @BeforeEach
+    private void beforeEach() {
+        answerSavedDeleteHistory = deleteHistoryRepository.save(DH_ANSWER);
+        questionDeleteHistory = deleteHistoryRepository.save(DH_QUESTION);
+    }
+
+
+    @Test
+    @DisplayName("delete history 등록")
+    public void saveDeleteHistoryTest() {
+        Long id = answerSavedDeleteHistory.getId();
+
+        Optional<DeleteHistory> oDeleteHistory = deleteHistoryRepository.findById(id);
+
+        assertAll(
+                () -> assertTrue(oDeleteHistory.isPresent()),
+                () -> assertEquals(oDeleteHistory.get(), answerSavedDeleteHistory)
+        );
+    }
+
+    @Test
+    @DisplayName("content type이 answer인 삭제 이력 조회")
+    public void findByContentTypeIsAnswerTest() {
+        Long answerId = answerSavedDeleteHistory.getId();
+        Long questionId = questionDeleteHistory.getId();
+
+        Optional<DeleteHistory> oAnswerHistory = deleteHistoryRepository.findById(answerId);
+        Optional<DeleteHistory> oQuestionHistory = deleteHistoryRepository.findById(questionId);
+
+        List<DeleteHistory> deleteHistories = deleteHistoryRepository.findByContentType(ContentType.ANSWER);
+
+        assertAll(
+                () -> assertThat(deleteHistories).contains(oAnswerHistory.get()),
+                () -> assertThat(deleteHistories).doesNotContain(oQuestionHistory.get())
+        );
+    }
+
+    @Test
+    @DisplayName("content type이 question인 삭제 이력 조회")
+    public void findByContentTypeIsQuestionTest() {
+        Long answerId = answerSavedDeleteHistory.getId();
+        Long questionId = questionDeleteHistory.getId();
+
+        Optional<DeleteHistory> oAnswerHistory = deleteHistoryRepository.findById(answerId);
+        Optional<DeleteHistory> oQuestionHistory = deleteHistoryRepository.findById(questionId);
+
+        List<DeleteHistory> deleteHistories = deleteHistoryRepository.findByContentType(ContentType.QUESTION);
+
+        assertAll(
+                () -> assertThat(deleteHistories).contains(oQuestionHistory.get()),
+                () -> assertThat(deleteHistories).doesNotContain(oAnswerHistory.get())
+        );
+    }
+}
